pyproject.toml
[tool.poetry]
name = "door-access-mgm"
version = "0.1.0"
description = ""
authors = ["Rodrigo de León <>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10, ^3.11"
django = "^5.0.4"
python-decouple = "^3.8"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "proj.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
